Move request util to utils directory
