+++ src/api/gin/Detail-funkcniho-mista.ts
@@ -1,7 +1,11 @@
 import type { Ginis } from '../../ginis'
 import { makeAxiosRequest } from '../../utils/api'
 import { GinisError } from '../../utils/errors'
-import { createXmlRequestBody, createXmlRequestConfig, extractResponseJson } from '../request-util'
+import {
+  createXmlRequestBody,
+  createXmlRequestConfig,
+  extractResponseJson,
+} from '../../utils/request-util'
 
 // https://robot.gordic.cz/xrg/Default.html?c=OpenMethodDetail&moduleName=SSL&version=390&methodName=Detail-funkcniho-mista&type=request
 const detailFunkcnihoMistaRequestProperties = ['Id-funkce'] as const
