+++ src/api/gin/Detail-referenta.ts
@@ -1,7 +1,11 @@
 import type { Ginis } from '../../ginis'
 import { makeAxiosRequest } from '../../utils/api'
 import { GinisError } from '../../utils/errors'
-import { createXmlRequestBody, createXmlRequestConfig, extractResponseJson } from '../request-util'
+import {
+  createXmlRequestBody,
+  createXmlRequestConfig,
+  extractResponseJson,
+} from '../../utils/request-util'
 
 // https://robot.gordic.cz/xrg/Default.html?c=OpenMethodDetail&moduleName=SSL&version=390&methodName=Detail-referenta&type=request
 const detailReferentaRequestProperties = ['Id-osoby'] as const
