+++ src/api/pod/Detail-el-podani.ts
@@ -1,7 +1,11 @@
 import type { Ginis } from '../../ginis'
 import { makeAxiosRequest } from '../../utils/api'
 import { GinisError } from '../../utils/errors'
-import { createXmlRequestBody, createXmlRequestConfig, extractResponseJson } from '../request-util'
+import {
+  createXmlRequestBody,
+  createXmlRequestConfig,
+  extractResponseJson,
+} from '../../utils/request-util'
 
 const detailElPodaniRequestProperties = [
   'Id-zpravy',
